[Android/Feat] 루틴 수정 기능 구현 S12P31D103-184

See merge request s12-final/S12P31D103!159

+++ Android/app/src/main/java/com/hogumiwarts/lumos/ui/common/DeviceGridHomeSection.kt
@@ -36,7 +36,7 @@ import com.hogumiwarts.lumos.ui.theme.nanum_square_neo
 @Composable
 fun DeviceGridHomeSection(
     devices: List<MyDevice>,
-    selectedDeviceId: Int? = null,
+    selectedDeviceId: Long? = null,
     onDeviceClick: (MyDevice) -> Unit = {},
     onToggleDevice: (MyDevice) -> Unit = {}
 ) {
