+++ Android/app/src/main/java/com/hogumiwarts/lumos/ui/screens/routine/routineCreate/SwipeableDeviceCardWithHint.kt
@@ -31,7 +31,7 @@ import kotlinx.coroutines.delay
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun SwipeableDeviceCardWithHint(
-    deviceId: Int,
+    deviceId: Long,
     shouldShowHint: Boolean,
     onHintShown: () -> Unit,
     onDelete: () -> Unit,
